新增存入excel方法,自然語言判斷研究中... 	new file:   Crawler/data.xlsx 	modified:   Crawler/main.py 	modified:   UserConfig/config.py 	modified:   test.py 	new file:   test.xlsx

+++ Crawler/main.py
@@ -6,25 +6,34 @@ from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.common.action_chains import ActionChains
-import time, json
+from openpyxl import Workbook, load_workbook
+import time, json, os
 
 #讀取設定檔
-with open('UserConfig/config.json', 'r', encoding='utf8')as J:     
+with open('UserConfig/config.json', 'r', encoding='utf8')as J:
     j = json.load(J)
 
-#預建立JSON資料檔案
-with open("Crawler/city.json", "a") as f:                           #city.json地區資料/job.json工作資料
-    json.dump([], f)
-    f.close()
-
 #寫入JSON
-def write_json(new_data, file='Crawler/city.json'):                 #city.json地區資料/job.json工作資料
+def write_json(new_data):                 
+    file = 'Crawler/city1.json'                                      #city.json地區資料/job.json工作資料
+    if not os.path.isfile(file):
+        with open(file, "w") as f:
+            json.dump([], f)
+            f.close()
+
     with open(file,'r+', encoding='utf8') as f:
         file_data = json.load(f)                                                #讀取JSON現有資料
         file_data.append(new_data)                                              #將新資料與舊資料結合
         f.seek(0)                                                               #設定輸入位置
         f.write(json.dumps(file_data, ensure_ascii = False, indent = 2))        #寫入                                                   
         f.close()                                                               #關閉
+
+'''存入excel才需啟用(9、80、89行也是)
+# wb = Workbook()                           #建立空白excel，最後存檔時才命名
+wb = load_workbook('data.xlsx')             #開啟excel
+wb.create_sheet("job")                      #建立工作表
+ws = wb['job']                              #選擇工作表
+'''
 
 #設定Chrome參數
 def setup():
@@ -45,7 +54,7 @@ WebDriverWait(driver, 100.0).until(
     )
 
 #模擬點擊[職務類別選單]
-ijob = driver.find_element(By.ID, "icity")                          #icity資料選單/ijob資料選單
+ijob = driver.find_element(By.ID, "ijob")                          #icity資料選單/ijob資料選單
 action = ActionChains(driver).click(ijob).perform()
 
 #等待網頁元素出現
@@ -69,7 +78,8 @@ for arrow in arrows:
         for l3 in l3s:
             if l3.text != '':
                 data.append(l3.text)
-        write_json({l2s[i].text: data,})                                             #呼叫寫入JSON方法
+                # ws.append([l3.text,l2s[i].text])                                     #寫入excel
+        write_json({l2s[i].text: data,})                                            #呼叫寫入JSON方法
         l3s.clear()                                                                  #清空l3s資料
         data.clear()                                                                 #清空data資料
         i += 1
@@ -77,4 +87,5 @@ for arrow in arrows:
         actions = ActionChains(driver).click(arrowup).perform()                     #關閉小類別
         time.sleep(0.3)                                                             #等待0.3秒
 
+# wb.save('data.xlsx')                                                                #excel存檔
 driver.quit()                                                                       #關閉網頁
